+++ lib/src/extensions.dart
@@ -18,6 +18,11 @@ extension RegexExtensionForStrings on String {
   ///
   /// Removes all non-numeric characters from the string.
   String get clean => replaceAll(RegExp(r'\s+'), '');
+
+  /// lastCharacters
+  ///
+  /// Returns the last [n] characters of the string.
+  String lastCharacters(int n) => substring(length - n);
 }
 
 /// RegexExtensionForNumbers
@@ -43,5 +48,34 @@ extension RegexExtensionForNumbers on String {
     }
 
     return number;
+  }
+}
+
+/// RegexExtensionForID
+///
+/// Extension methods for the [String] class.
+extension RegexExtensionForID on String {
+  /// formatID
+  ///
+  /// Returns string in the requested format.
+  /// @throws [LocalRegexException] if the national id is invalid.
+  String? formatID() {
+    late String id;
+    if (LocalRegex.isZimID(this)) {
+      final idWithNoSpaces = clean.replaceAll(RegExp('[^A-Z0-9]'), '');
+      final idPrefix = idWithNoSpaces.substring(0, 2);
+      final idSuffix = idWithNoSpaces.lastCharacters(3);
+      final properIdSuffix =
+          '${idSuffix.split('')[0]} ${idSuffix.split('')[1]}${idSuffix.split('')[2]}';
+      final idBody = idWithNoSpaces
+          .replaceAll(idPrefix, '')
+          .replaceAll(idSuffix, '')
+          .trim();
+      id = '$idPrefix-$idBody $properIdSuffix';
+    } else {
+      throw LocalRegexException('National ID is not valid');
+    }
+
+    return id;
   }
 }